Move containers to container subdir.
